+++ src/main/java/roomescape/controller/UserPageController.java
@@ -4,27 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class PageController {
-
-    @GetMapping("/admin")
-    public String admin() {
-        return "/admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String getReservation() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/admin/theme")
-    public String getTheme() {
-        return "admin/theme";
-    }
-
-    @GetMapping("/admin/time")
-    public String time() {
-        return "admin/time";
-    }
+public class UserPageController {
 
     @GetMapping("/reservation")
     public String reservation() {
