+++ src/test/java/roomescape/controller/AdminPageControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.MediaType;
 import roomescape.dto.LoginRequestDto;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class PageControllerTest {
+public class AdminPageControllerTest {
 
     String accessToken;
 
@@ -66,33 +66,6 @@ public class PageControllerTest {
         RestAssured.given().log().all()
                 .cookie("token", accessToken)
                 .when().get("/admin/time")
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    @DisplayName("유저 예약 페이지 요청 시 200으로 응답한다.")
-    @Test
-    void welcomePageTest() {
-        RestAssured.given().log().all()
-                .when().get("/reservation")
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    @DisplayName("인기 테마 페이지 요청 시 200으로 응답한다.")
-    @Test
-    void bestPageTest() {
-        RestAssured.given().log().all()
-                .when().get("/")
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    @DisplayName("로그인 페이지 요청 시 200으로 응답한다.")
-    @Test
-    void loginPageTest() {
-        RestAssured.given().log().all()
-                .when().get("/login")
                 .then().log().all()
                 .statusCode(200);
     }
