+++ src/test/java/roomescape/controller/AdminReservationThemeControllerTest.java
@@ -20,7 +20,7 @@ import static org.hamcrest.Matchers.is;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @Sql(scripts = {"/test_schema.sql", "/test_data.sql"})
-public class ReservationThemeControllerTest {
+public class AdminReservationThemeControllerTest {
 
     @LocalServerPort
     int port;
@@ -94,15 +94,5 @@ public class ReservationThemeControllerTest {
                 .then().log().all()
                 .statusCode(400)
                 .body(is("예약이 존재하는 테마는 삭제할 수 없습니다."));
-    }
-
-    @DisplayName("최근 일주일간 가장 예약이 많은 테마를 조회한다.")
-    @Test
-    void weeklyBestThemesTest() {
-        RestAssured.given().log().all()
-                .when().get("/best-themes")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(3));
     }
 }
