+++ src/test/java/roomescape/controller/UserPageControllerTest.java
@@ -1,0 +1,47 @@
+package roomescape.controller;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class UserPageControllerTest {
+
+    @LocalServerPort
+    int port;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
+
+    @DisplayName("유저 예약 페이지 요청 시 200으로 응답한다.")
+    @Test
+    void welcomePageTest() {
+        RestAssured.given().log().all()
+                .when().get("/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("인기 테마 페이지 요청 시 200으로 응답한다.")
+    @Test
+    void bestPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("로그인 페이지 요청 시 200으로 응답한다.")
+    @Test
+    void loginPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/login")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
